Split out some logic that will need to be reused as utils
